Continued database work, finished Insert methods

+++ Handlers/ProfileHandler/Profile.cs
@@ -7,7 +7,7 @@ using Discord;
 
 namespace _490Bot.Handlers.ProfileHandler {
     public class Profile {
-        public int UserID { get; set; }
+        public ulong UserID { get; set; }
         public String StatusField { get; set; }
         public String AboutField { get; set; }
         public Badge[] Badges { get; set; }
